+++ src/GOODFFrame.cpp
@@ -47,6 +47,7 @@ BEGIN_EVENT_TABLE(GOODFFrame, wxFrame)
 	EVT_MENU(ID_IMPORT_VOICING_DATA, GOODFFrame::OnImportCMB)
 	EVT_MENU(ID_IMPORT_STOP_RANK, GOODFFrame::OnImportStopRank)
 	EVT_MENU(ID_GLOBAL_SHOW_TOOLTIPS_OPTION, GOODFFrame::OnEnableTooltipsMenu)
+	EVT_MENU(ID_GLOBAL_PARSE_LEGACY_XFADES_OPTION, GOODFFrame::OnParseLegacyXfadesMenu)
 	EVT_MENU(ID_CLEAR_HISTORY, GOODFFrame::OnClearHistory)
 	EVT_MENU(ID_DEFAULT_PATHS_MENU, GOODFFrame::OnDefaultPathMenuChoice)
 	EVT_MENU_RANGE(wxID_FILE1, wxID_FILE9, GOODFFrame::OnRecentFileMenuChoice)
@@ -77,6 +78,9 @@ GOODFFrame::GOODFFrame(const wxString& title) : wxFrame(NULL, wxID_ANY, title) {
 	m_config = new wxFileConfig(wxT("GoOdf"));
 	m_defaultOrganDirectory = wxEmptyString;
 	m_defaultCmbDirectory = wxEmptyString;
+	m_parseLegacyXfades = false;
+	m_logWindow = new wxLogWindow(this, wxT("Log messages"), false, false);
+	wxLog::SetActiveTarget(m_logWindow);
 
 	// Create a file menu
 	m_fileMenu = new wxMenu();
@@ -98,6 +102,8 @@ GOODFFrame::GOODFFrame(const wxString& title) : wxFrame(NULL, wxID_ANY, title) {
 	m_toolsMenu->Append(ID_IMPORT_STOP_RANK, wxT("Import Stops/Ranks\tCtrl+R"), wxT("Import stops/ranks from another (working) .organ file"));
 	m_toolsMenu->AppendCheckItem(ID_GLOBAL_SHOW_TOOLTIPS_OPTION, wxT("Enable Tooltips"), wxT("Enable tooltips for certain controls"));
 	m_toolsMenu->Check(ID_GLOBAL_SHOW_TOOLTIPS_OPTION, false);
+	m_toolsMenu->AppendCheckItem(ID_GLOBAL_PARSE_LEGACY_XFADES_OPTION, wxT("Parse Legacy X-fades"), wxT("When opening an .organ file, make attacks/releases inherit LoopCrossfadeLength & ReleaseCrossfadeLength values like pre GO v3.14.0"));
+	m_toolsMenu->Check(ID_GLOBAL_PARSE_LEGACY_XFADES_OPTION, false);
 	m_toolsMenu->Append(ID_CLEAR_HISTORY, wxT("Clear File History"), wxT("Remove all the entries in the recent file history"));
 	m_toolsMenu->Append(ID_DEFAULT_PATHS_MENU, wxT("Default paths\tCtrl+P"), wxT("Set the default paths used by the application here"));
 
@@ -474,6 +480,8 @@ GOODFFrame::GOODFFrame(const wxString& title) : wxFrame(NULL, wxID_ANY, title) {
 GOODFFrame::~GOODFFrame() {
 	if (m_organ)
 		delete m_organ;
+	wxLog::SetActiveTarget(nullptr);
+	delete m_logWindow;
 }
 
 void GOODFFrame::OnAbout(wxCommandEvent& WXUNUSED(event)) {
@@ -615,6 +623,7 @@ void GOODFFrame::DoOpenOrgan(wxString filePath) {
 	m_organ = new Organ();
 	removeAllItemsFromTree();
 	m_organHasBeenSaved = false;
+	RecreateLogWindow();
 
 	OrganFileParser parser(filePath, m_organ);
 	if (parser.isOrganReady()) {
@@ -680,6 +689,8 @@ void GOODFFrame::DoOpenOrgan(wxString filePath) {
 			RebuildPanelGuiElementsInTree(i);
 		}
 		UpdateFrameTitle();
+		if (m_parseLegacyXfades)
+			m_organ->doInheritLegacyXfades();
 	} else {
 		if (m_organ) {
 			delete m_organ;
@@ -2707,6 +2718,7 @@ void GOODFFrame::OnNewOrgan(wxCommandEvent& WXUNUSED(event)) {
 			m_organPanel->setOdfPath(wxEmptyString);
 			m_organPanel->setOdfName(wxEmptyString);
 			SetTitle(::wxGetApp().m_fullAppName);
+			RecreateLogWindow();
 		}
 	} else {
 		if (m_organ) {
@@ -2724,6 +2736,7 @@ void GOODFFrame::OnNewOrgan(wxCommandEvent& WXUNUSED(event)) {
 		m_organPanel->setOdfPath(wxEmptyString);
 		m_organPanel->setOdfName(wxEmptyString);
 		SetTitle(::wxGetApp().m_fullAppName);
+		RecreateLogWindow();
 	}
 }
 
@@ -3151,6 +3164,15 @@ void GOODFFrame::OnImportStopRank(wxCommandEvent& WXUNUSED(event)) {
 	delete sourceOrgan;
 }
 
+void GOODFFrame::OnParseLegacyXfadesMenu(wxCommandEvent& event) {
+	// ID_GLOBAL_PARSE_LEGACY_XFADES_OPTION
+	if (m_toolsMenu->IsChecked(ID_GLOBAL_PARSE_LEGACY_XFADES_OPTION)) {
+		m_parseLegacyXfades = true;
+	} else {
+		m_parseLegacyXfades = false;
+	}
+}
+
 void GOODFFrame::SetupOrganMainPanel() {
 	// Main panel is created by Organ itself but we need to add it to the tree
 	wxTreeItemId mainPanel = m_organTreeCtrl->AppendItem(tree_panels, m_organ->getOrganPanelAt(0)->getName());
@@ -3246,6 +3268,14 @@ wxString GOODFFrame::GetDefaultCmbDirectory() {
 	return m_defaultCmbDirectory;
 }
 
+wxLogWindow* GOODFFrame::GetLogWindow() {
+	return m_logWindow;
+}
+
+bool GOODFFrame::IsParsingLegacyXfades() {
+	return m_parseLegacyXfades;
+}
+
 void GOODFFrame::removeAllItemsFromTree() {
 	m_organTreeCtrl->DeleteChildren(tree_manuals);
 	m_organTreeCtrl->DeleteChildren(tree_windchestgrps);
@@ -3267,4 +3297,11 @@ void GOODFFrame::UpdateFrameSizeAndPos() {
 		GetSize(&m_frameWidth, &m_frameHeight);
 		m_frameMaximized = false;
 	}
+}
+
+void GOODFFrame::RecreateLogWindow() {
+	wxLog::SetActiveTarget(nullptr);
+	delete m_logWindow;
+	m_logWindow = new wxLogWindow(this, wxT("Log messages"), false, false);
+	wxLog::SetActiveTarget(m_logWindow);
 }