+++ src/GOODFFrame.h
@@ -80,6 +80,8 @@ public:
 	void SynchronizePipeReadingOptions(RankPanel* rankPanel, wxString atkFolder, bool oneAttack, bool loadRelease, wxString releaseFolder, bool extractTime, wxString tremFolder);
 	wxString GetDefaultOrganDirectory();
 	wxString GetDefaultCmbDirectory();
+	wxLogWindow* GetLogWindow();
+	bool IsParsingLegacyXfades();
 
 	Organ *m_organ;
 
@@ -125,6 +127,7 @@ private:
 	wxTreeItemId tree_generals;
 	wxTreeItemId tree_panels;
 	wxTreeItemId m_draggedItem;
+	wxLogWindow *m_logWindow;
 
 	OrganPanel *m_organPanel;
 	EnclosurePanel *m_enclosurePanel;
@@ -158,6 +161,7 @@ private:
 	bool m_frameMaximized;
 	wxString m_defaultOrganDirectory;
 	wxString m_defaultCmbDirectory;
+	bool m_parseLegacyXfades;
 
 	void OnOrganTreeSelectionChanged(wxTreeEvent& event);
 	void OnOrganTreeRightClicked(wxTreeEvent& event);
@@ -182,10 +186,12 @@ private:
 	void OnSizeChange(wxSizeEvent& event);
 	void OnDefaultPathMenuChoice(wxCommandEvent& event);
 	void OnImportStopRank(wxCommandEvent& event);
+	void OnParseLegacyXfadesMenu(wxCommandEvent& event);
 
 	void SetupOrganMainPanel();
 	void removeAllItemsFromTree();
 	void UpdateFrameSizeAndPos();
+	void RecreateLogWindow();
 
 };